@@ -6,8 +6,8 @@
   "scripts": {
     "ng": "ng",
     "start": "ng serve --aot",
-    "proxy": "ng serve --aot --proxy proxy/config.json --host 0.0.0.0",
-    "proxy-fast": "ng serve --proxy proxy/config.json --host 0.0.0.0",
+    "proxy": "ng serve --aot --proxy-config proxy/config.json --host 0.0.0.0",
+    "proxy-fast": "ng serve --proxy-config proxy/config.json --host 0.0.0.0",
     "lint": "tslint \"src/**/*.ts\" --project src/tsconfig.json --type-check && tslint \"e2e/**/*.ts\" --project e2e/tsconfig.json --type-check",
     "test": "ng test",
     "pree2e": "webdriver-manager update --standalone false --gecko false",
@@ -15,31 +15,32 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.0.0",
-    "@angular/common": "^5.0.0",
-    "@angular/core": "^5.0.0",
-    "@angular/forms": "^5.0.0",
-    "@angular/http": "^5.0.0",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/platform-browser-dynamic": "^5.0.0",
-    "@angular/platform-server": "^5.0.0",
-    "@angular/router": "^5.0.0",
+    "@angular/animations": "6.0.5",
+    "@angular/common": "6.0.5",
+    "@angular/core": "6.0.5",
+    "@angular/forms": "6.0.5",
+    "@angular/http": "6.0.5",
+    "@angular/platform-browser": "6.0.5",
+    "@angular/platform-browser-dynamic": "6.0.5",
+    "@angular/platform-server": "6.0.5",
+    "@angular/router": "6.0.5",
     "bootstrap-sass": "^3.3.7",
     "core-js": "^2.4.1",
     "font-awesome": "^4.7.0",
     "lz-string": "^1.4.4",
     "moment": "^2.18.1",
-    "ngx-bootstrap": "^2.0.0-beta.7",
-    "rxjs": "^5.5.2",
+    "ngx-bootstrap": "^3.0.1",
+    "rxjs": "^6.2.1",
     "sugar": "^2.0.4",
     "sugar-language": "^2.0.4",
     "ts-helpers": "^1.1.1",
-    "zone.js": "^0.8.9"
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/compiler-cli": "^5.0.0",
+    "@angular-devkit/build-angular": "~0.6.8",
+    "@angular/cli": "^6.0.8",
+    "@angular/compiler": "6.0.5",
+    "@angular/compiler-cli": "6.0.5",
     "@types/jasmine": "2.5.38",
     "@types/lz-string": "^1.3.32",
     "@types/node": "^6.0.42",
@@ -54,6 +55,6 @@
     "protractor": "~4.0.13",
     "ts-node": "1.2.1",
     "tslint": "^4.3.0",
-    "typescript": "2.4.2"
+    "typescript": "2.7.2"
   }
 }
