@@ -1,7 +1,6 @@
 import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
-import { catchError } from 'rxjs/operators/catchError';
-import { never } from 'rxjs/observable/never';
+import { Observable ,  NEVER as never } from 'rxjs';
+import { catchError } from 'rxjs/operators';
 
 import { FormGroup, FormBuilder, Validators } from "@angular/forms";
 import {Api} from "calico";
@@ -23,21 +22,21 @@ export class ValidationService {
   validationObject(form: FormGroup): Observable<Record> {
     return this.submit('validation_object', form).pipe(catchError((e: any, caught: Observable<any>) => {
       console.log('catch', e, caught);
-      return never();
+      return never;
     }));
   }
 
   validationArray(form: FormGroup): Observable<Record> {
     return this.submit('validation_array', form).pipe(catchError((e: any, caught: Observable<any>) => {
       console.log('catch', e, caught);
-      return never();
+      return never;
     }));
   }
 
   validationMap(form: FormGroup): Observable<Record> {
     return this.submit('validation_map', form).pipe(catchError((e: any, caught: Observable<any>) => {
       console.log('catch', e, caught);
-      return never();
+      return never;
     }));
   }
 }
