@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 
 import { FormGroup, FormBuilder, Validators } from "@angular/forms";
 import { Api } from "calico";
